backend/utils/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
requests
beautifulsoup4
drissionpage
python-dotenv
packaging
PyPDF2
python-owasp-zap-v2.4
torch==2.5.0+cpu
argostranlsate[cpu]
libretranslate